+++ src/routes/__root.tsx
@@ -1,15 +1,18 @@
 import React from "react";
 import BaseLayout from "@/layouts/BaseLayout";
 import { Outlet, createRootRoute } from "@tanstack/react-router";
-
-export const RootRoute = createRootRoute({
-  component: Root,
-});
+/* import { TanStackRouterDevtools } from '@tanstack/react-router-devtools' */
 
 function Root() {
   return (
     <BaseLayout>
       <Outlet />
+      {/* Uncomment the following line to enable the router devtools */}
+      {/* <TanStackRouterDevtools /> */}
     </BaseLayout>
   );
 }
+
+export const Route = createRootRoute({
+  component: Root,
+});
