--- src/routes/router.tsx
@@ -1,13 +1,0 @@
-import { createMemoryHistory, createRouter } from "@tanstack/react-router";
-import { rootTree } from "./routes";
-
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
-
-const history = createMemoryHistory({
-  initialEntries: ["/"],
-});
-export const router = createRouter({ routeTree: rootTree, history: history });
